+++ src/UiAdminBase/Contract/Action/PortalNode/PortalNodeAddUiActionInterface.php
@@ -6,11 +6,13 @@ namespace Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Action\PortalNode;
 
 use Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeAdd\PortalNodeAddPayload;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Action\PortalNode\PortalNodeAdd\PortalNodeAddResult;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Action\UiActionContextInterface;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Action\UiActionInterface;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\InvalidArgumentThrowableInterface;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PersistException;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PortalNodeAliasIsAlreadyAssignedException;
 
-interface PortalNodeAddUiActionInterface
+interface PortalNodeAddUiActionInterface extends UiActionInterface
 {
     /**
      * Create a new portal node by given portal class and alias.
@@ -19,5 +21,5 @@ interface PortalNodeAddUiActionInterface
      * @throws PersistException
      * @throws PortalNodeAliasIsAlreadyAssignedException
      */
-    public function add(PortalNodeAddPayload $payload): PortalNodeAddResult;
+    public function add(PortalNodeAddPayload $payload, UiActionContextInterface $context): PortalNodeAddResult;
 }