+++ src/UiAdminBase/Contract/Action/Route/RouteAddUiActionInterface.php
@@ -6,13 +6,15 @@ namespace Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Action\Route;
 
 use Heptacom\HeptaConnect\Ui\Admin\Base\Action\Route\RouteAdd\RouteAddPayloadCollection;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Action\Route\RouteAdd\RouteAddResultCollection;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Action\UiActionContextInterface;
+use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Action\UiActionInterface;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\InvalidArgumentThrowableInterface;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PersistException;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\PortalNodesMissingException;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\RouteAddFailedException;
 use Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Exception\RouteAlreadyExistsException;
 
-interface RouteAddUiActionInterface
+interface RouteAddUiActionInterface extends UiActionInterface
 {
     /**
      * Creates new routes between portal nodes by given entities and capabilities.
@@ -23,5 +25,5 @@ interface RouteAddUiActionInterface
      * @throws RouteAddFailedException
      * @throws RouteAlreadyExistsException
      */
-    public function add(RouteAddPayloadCollection $payloads): RouteAddResultCollection;
+    public function add(RouteAddPayloadCollection $payloads, UiActionContextInterface $context): RouteAddResultCollection;
 }